Moved Sync Functions to Typescript
